+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased changes
+## 0.13.0
 
 - Add `t:Pages.result/0` to reflect that any function may return an error.
 - Pages return `{:error, :external, url}` when given a non-local URL.